+++ src/main/java/com/imsweb/geocoder/component/SummaryPanel.java
@@ -104,6 +104,15 @@ public class SummaryPanel extends JPanel {
         constraints.gridy++;
         constraints.gridx = 0;
         constraints.anchor = GridBagConstraints.EAST;
+        summaryPnl.add(Utils.createLabel("number rejected: "), constraints);
+        constraints.gridx = 1;
+        constraints.anchor = GridBagConstraints.WEST;
+        percentage = (float)_parent.getSession().getNumRejectedLines() / numResultLines * 100;
+        summaryPnl.add(new JLabel(Integer.toString(session.getNumRejectedLines()) + " (" + String.format("%.1f", percentage) + "%)"), constraints);
+
+        constraints.gridy++;
+        constraints.gridx = 0;
+        constraints.anchor = GridBagConstraints.EAST;
         summaryPnl.add(Utils.createLabel("number skipped: "), constraints);
         constraints.gridx = 1;
         constraints.anchor = GridBagConstraints.WEST;