+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## NAACCR Geocoder Review Version History
 
-**Version 1.0**
+**Version 0.2**
+
+ - Added support for rejecting all geocoder results.
+ - Improved error messages when the JSON cannot be properly parsed.
+ - Added NAACCR icons.
+
+**Version 0.1**
 
  - Initial version.